bug 1171957 - Renaming specification module
It makes sense for the visitors for SpecName, Spec2, and the
specification description should be in the same module.
